Don't register zero compromise times when not needed.
Mike noticed our "time to G2" graphs have higher probability mass in the 0
hours mark. A bug that might have contributed to that is that we were assigning
0-hour compromise times to G2 and G3, when a Sybil was successful against G1
during experiment initialization (that has a non-negligible probability of
happening).

We are still assigning 0-hour compromise times, if a Sybil succeeds against G2
or G3 during experiment initialization.